fixup relative_energy interaction with eas_deadzone_filter
when schedtune is in use, we have a percentage in the delta member of
energy_env for each target CPU. We cannot use a scaling of the prev
energy as the margin, so here we switch to using a set 3% margin (as
we used to use with the >>6).

Normalized energy needs something a bit different, but I haven't gone
through the math to figure it out yet.

Also, I haven't touched the schedtune filtering which I think is also
needing an update to handle the relative energy diff.

Change-Id: I1af26f37596ae6040625e754b1af5a88c4ef8145
Signed-off-by: Chris Redpath <chris.redpath@arm.com>
(cherry picked from commit c680ed127b7a4f8dc411045ea21c0ba151dfcd89)
(cherry picked from commit 965edef93340e5518a5d960bd28d6c7144dd6c9f)